added jwt validation for plan requests

+++ api/controllers/plans.js
@@ -2,12 +2,23 @@
 const db = require("../services/plans");
 const etag = require("etag");
 const {payloadValidation} = require("../helper/schemaValidator");
+const {generateToken,validateToken} = require("../middleware/auth");
 
 
 module.exports = {
 
     getPlanById : async(req,res) => {
         try {
+            var token = req.header("x-auth-token");
+            if(!token){
+                return res.status(401).json({msg : 'No token, authorization denied'});
+           }
+            result = validateToken(token);
+            console.log("check 4");
+            if (!result) {
+                console.log("check 6");
+                return res.status(400).json({"msg":"token invalidate"});
+            } 
             if(req.params.id == null){
                 return res.status(400).json({
                     "msg":"bad request"
@@ -44,7 +55,18 @@ module.exports = {
     },
     savePlan : async(req,res) => {
         try {
+            var token = req.header("x-auth-token");
+            if(!token){
+                return res.status(401).json({msg : 'No token, authorization denied'});
+           }
+            result = validateToken(token);
+            console.log("check 4");
+            if (!result) {
+                console.log("check 6");
+                return res.status(400).json({"msg":"token invalidate"});
+            } 
             if (payloadValidation(req)) {
+                
                 console.log(req.body);
                 console.log("check 0",req.body.objectId);
                 const payload = await db.getPlanById(req.body.objectId)
@@ -74,7 +96,18 @@ module.exports = {
     },
     deletePlanById : async (req,res) => {
         try {
+            var token = req.header("x-auth-token");
+            if(!token){
+                return res.status(401).json({msg : 'No token, authorization denied'});
+           }
+            result = validateToken(token);
+            console.log("check 4");
+            if (!result) {
+                console.log("check 6");
+                return res.status(400).json({"msg":"token invalidate"});
+            } 
             const payload = await db.getPlanById(req.params.id)
+
             if (payload){
                 const del = await db.deletePlanById(req.params.id)
                 if(del == true){
